ai_services/training/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Deep Learning Frameworks
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
deepspeed>=0.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
noisereduce>=2.0.0
pydub>=0.25.0

# NLP and Text Processing
tokenizers>=0.13.0
sentence-transformers>=2.2.0
nltk>=3.8.0
spacy>=3.6.0
langdetect>=1.0.9

# Data Handling
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
h5py>=3.9.0
lmdb>=1.4.0

# Image Processing (for evaluation)
Pillow>=9.5.0
opencv-python>=4.7.0

# Logging and Visualization
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Model Checkpointing
safetensors>=0.3.0

# Distributed Training
mpi4py>=3.1.0

# Quality Assurance
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utility
tqdm>=4.65.0
regex>=2023.0.0
joblib>=1.3.0
